requirements.txt
jittor==1.3.4.13
Pillow==9.1.0
numpy==1.22.3
scipy==1.8.0
imageio-ffmpeg==0.4.7 
imageio==2.19.1
opencv-python==4.5.5.64
scikit-image==0.19.2
dlib==19.23.1
tqdm==4.63.0
PyYAML==6.0
cupy-cuda111
